requirements.txt
aws-cdk-lib==2.10.0
constructs>=10.0.0,<11.0.0
cdk-ecr-deployment==2.1.3
#cdk-datadog-integration==1.3.2
#cdk-dia==0.6.0
cdk8s==2.2.9